refactor: Interest, UserInterest 패키지 분리
